improve logic to use util, move corecheck in /cluster
