+++ pkg/collector/corechecks/cluster/doc.go
@@ -1,0 +1,10 @@
+// Unless explicitly stated otherwise all files in this repository are licensed
+// under the Apache License Version 2.0.
+// This product includes software developed at Datadog (https://www.datadoghq.com/).
+// Copyright 2017 Datadog, Inc.
+
+/*
+Package containers provides core checks for cluster level checks, used by the Datadog Cluster Agent.
+
+*/
+package cluster
